requirements.txt
requests
PyQt5
sphinx
python-dateutil
